Test exit code of filesystem check is 4
Previously, we was checking it manually, so the CI would always succeed
regardless if the expected behavior was demonstrated. We can automate
this check by checking the exit code of the filesystem check.  If the
exit code is 4, the filesystem check failed as expected. However, it
doesn't contain any fatal errors that could cause failures.