Add version to caches
Since we setup E2E tests environment from scratch every time, it's not
necessary. But when we run tests in non-E2E environment, we need to add
version to caches to avoid conflicts.

Also if the both runner type are running in the same time, their caches
also will be conflicted in E2E environment too.